oops wrong idea

+++ mgx_scripts/iterative_growth_and_measures.py
@@ -133,15 +133,19 @@ def step_check(path, filename):
     '''
     step_file = os.path.join(path, filename)
 
-    with open(step_file, 'r') as f:
-        last_step = f.read()
-        if last_step == "":
-            step = 1
-        else:
-            step = int(last_step) + 1
-    with open(step_file, 'w') as f:
-        f.write(str(step))
+    step = 1
 
+    if os.path.exists(step_file):
+        with open(step_file, 'r') as f:
+            last_step = f.read()
+            step = int(last_step) + 1
+        with open(step_file, 'w') as f:
+            f.write(str(step))
+
+
+    else:
+        with open(step_file, 'w') as f:
+            f.write(str(step))
 
     return step
